configuration parameters moved to config.ts
