+++ src/app/config.ts
@@ -1,0 +1,38 @@
+import { Configuration } from "msal";
+
+export const isIE = window.navigator.userAgent.indexOf('MSIE ') > -1 || window.navigator.userAgent.indexOf('Trident/') > -1;
+
+// Config object to be passed to Msal on creation.
+// For a full list of msal.js configuration parameters, 
+// visit https://azuread.github.io/microsoft-authentication-library-for-js/docs/msal/modules/_configuration_.html
+export const msalConfig: Configuration = {
+    auth: {
+        clientId: 'e760cab2-b9a1-4c0d-86fb-ff7084abd902', 
+        authority: 'https://fabrikamb2c.b2clogin.com/fabrikamb2c.onmicrosoft.com/b2c_1_susi',
+        validateAuthority: false
+    },
+    cache: {
+        cacheLocation: 'localStorage', // This configures where your cache will be stored
+        storeAuthStateInCookie: isIE, // Set this to "true" to save cache in cookies to address trusted zones limitations in IE (see: https://github.com/AzureAD/microsoft-authentication-library-for-js/wiki/Known-issues-on-IE-and-Edge-Browser)
+    },
+}
+
+// main API configuration parameters
+export const apiConfig = {
+    b2cScopes: ["https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read"],
+    webApi: "https://fabrikamb2chello.azurewebsites.net/hello"
+};
+
+// Add here scopes for id token
+// For a full list of available authentication parameters, 
+// visit https://azuread.github.io/microsoft-authentication-library-for-js/docs/msal/modules/_authenticationparameters_.html
+export const loginRequest = {
+    scopes: ["openid", "profile"],
+};
+
+// Add here scopes for access token to be used at the API endpoints.
+export const tokenRequest = {
+    scopes: ["https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read"]
+};
+
+
