+++ src/environments/environment.ts
@@ -4,28 +4,10 @@
 
 export const environment = {
   production: false,
-  isIE: false,
-  msalConfig: {
-    auth: {
-        clientId: 'e760cab2-b9a1-4c0d-86fb-ff7084abd902', 
-        authority: 'https://fabrikamb2c.b2clogin.com/fabrikamb2c.onmicrosoft.com/b2c_1_susi',
-        validateAuthority: false
-    },
-    cache: {
-        cacheLocation: 'localStorage',
-        storeAuthStateInCookie: false,
-    },
-  },
-  apiConfig: {
-    b2cScopes: ["https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read"],
-    webApi: "https://fabrikamb2chello.azurewebsites.net/hello"
-  },
-  loginRequest: {
-    scopes: ["openid", "profile"],
-  },
-  tokenRequest: {
-    scopes: ["https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read"]
-  },
+  clientId: 'e760cab2-b9a1-4c0d-86fb-ff7084abd902',
+  authority: 'https://fabrikamb2c.b2clogin.com/fabrikamb2c.onmicrosoft.com/b2c_1_susi',
+  webApi: "https://fabrikamb2chello.azurewebsites.net/hello",
+  b2cScopes: ["https://fabrikamb2c.onmicrosoft.com/helloapi/demo.read"],
 };
 
 /*
